Add test case illustrating slow planning
Flows that involve many CoGroups can take a very long time to
plan. This is a test case that shows a flow that can cause
this problem. This is a pretty contrived flow, but we have seen
this issue with other more realistic flows. I only set the flow
up like this (chaining together n pipes using CoGroups) because
it's easy to measure the growth in running time as we add more
pipes.